Shiny Application that calculates PIF or PAF
Shiny Application that calculates PIF or PAF as chosen by the user. User can upload values for exposure levels and weights, and either choose or upload the relative risk function as well as the counterfactual function. The option where the minimum exposure is different to zero isn't working yet. The Shiny Application requires the package PIF to work.